REFACTORING CODE so that same files are at the same place
