+++ app/src/main/java/com/example/tryggaklassenpod/screens/PodcastViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.tryggaklassenpod.screens
+
+import androidx.lifecycle.ViewModel
+
+class PodcastViewModel(): ViewModel() {
+
+}
