--- app/src/main/java/com/example/tryggaklassenpod/ui/screen/playerScreen/PlayerViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.tryggaklassenpod.ui.screen.playerScreen
-
-import androidx.lifecycle.ViewModel
-
-class PlayerViewModel(): ViewModel() {
-
-}
